backend/requirements.txt
beautifulsoup4==4.14.3
et_xmlfile==2.0.0
greenlet==3.3.0
numpy==2.4.1
openpyxl==3.1.5
pandas==2.3.3
playwright==1.57.0
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
soupsieve==2.8.1
typing_extensions==4.15.0
tzdata==2025.3
